refactor tests to use conftest.py, which contains global test data
