+++ tests/test_model.py
@@ -1,23 +1,17 @@
 import pytest
 import torch
 
-from argus.models import NCameraCNN, NCameraCNNConfig
+from argus.models import NCameraCNN
 
 
-@pytest.fixture(scope="session")
-def model() -> NCameraCNN:
-    """A fixture for the model."""
-    return NCameraCNN(NCameraCNNConfig(n_cams=2, W=672, H=376))
-
-
-def test_forward(model: NCameraCNN) -> None:
+def test_forward(dummy_model: NCameraCNN) -> None:
     """Tests the forward pass."""
     # unbatched, assert throws error
     x = torch.randn(6, 376, 672)
     with pytest.raises(AssertionError):
-        pose = model(x)
+        pose = dummy_model(x)
 
     # batch
     x = torch.randn(2, 6, 376, 672)
-    pose = model(x)
+    pose = dummy_model(x)
     assert pose.shape == (2, 6)
